Allow shell to changed via config.
This commit allows me (and others!) to change which shell to use to run the tests via the `config`. This is useful for platforms that do not have the default shell, `bash`, present.

To test this I decided that if the `lit` crate is happy to rely on `bash`, it makes sense that `sh` would also be present. So I added an additional run of the integration tests using `sh` to prove that it works.